+++ src/OpenTelemetry.OpAmp.Client/CHANGELOG.md
@@ -8,6 +8,9 @@
 * Add support for reporting full state.
   ([#4633](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4633))
 
+* Add heartbeat support to full state report.
+  ([#4653](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4653))
+
 ## 0.5.0-alpha.1
 
 Released 2026-Jul-01